Programming lab5 workversion 0.1 without script

+++ Программирование/Lab5/src/main/java/common/Model/Location.java
@@ -18,7 +18,7 @@ public class Location implements Comparable<Location>{
         z = zcoord;
     }
 
-    public float getX() {
+    public int getX() {
         return x;
     }
